# [SPARK-38867][SQL] Avoid OOM when bufferedPlan has a lot of duplicate keys in SortMergeJoin codegen 