--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -2490,6 +2490,15 @@ object SQLConf {
       .intConf
       .createWithDefault(SHUFFLE_SPILL_NUM_ELEMENTS_FORCE_SPILL_THRESHOLD.defaultValue.get)
 
+  val SORT_MERGE_JOIN_EXEC_MAX_RECORD_PER_CYCLE =
+    buildConf("spark.sql.sortMergeJoinExec.codegen.maxRecordPerCycle")
+      .internal()
+      .doc("Every time so many records are output, check whether SortMergeJoin " +
+        "should be stopped, and then trigger parent to consume to avoid OOM.")
+      .version("3.4.0")
+      .intConf
+      .createWithDefault(1000)
+
   val CARTESIAN_PRODUCT_EXEC_BUFFER_IN_MEMORY_THRESHOLD =
     buildConf("spark.sql.cartesianProductExec.buffer.in.memory.threshold")
       .internal()
@@ -4301,6 +4310,9 @@ class SQLConf extends Serializable with Logging {
 
   def sortMergeJoinExecBufferSpillThreshold: Int =
     getConf(SORT_MERGE_JOIN_EXEC_BUFFER_SPILL_THRESHOLD)
+
+  def sortMergeJoinExecMaxRecordPerCycle: Int =
+    getConf(SORT_MERGE_JOIN_EXEC_MAX_RECORD_PER_CYCLE)
 
   def cartesianProductExecBufferInMemoryThreshold: Int =
     getConf(CARTESIAN_PRODUCT_EXEC_BUFFER_IN_MEMORY_THRESHOLD)