@cloud-fan 2022-04-18
Can we run some benchmarks to evaluate the perf impact? This adds a breakpoint to the SMJ workflow which may slow down the join processing.