@cxzl25 2022-04-21
Encountered the same problem in production, but I simply added a configuration to disable codegen for SMJ.