@mcdull-zhang 2022-04-24
> Can we run some benchmarks to evaluate the perf impact? This adds a breakpoint to the SMJ workflow which may slow down the join processing.

The benchmark test has been updated to the description of pr. Could you please take a look at this?